refactor: Split models into directories
This'll allow adding supporting files to each model.
