+++ src/decoder/CMakeLists.txt
@@ -1,0 +1,8 @@
+target_sources(tts PRIVATE
+        dac_model.cpp
+        dac_model.h
+        general_neural_audio_codec.cpp
+        general_neural_audio_codec.h
+        snac_model.cpp
+        snac_model.h
+)
