+++ src/models/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_subdirectory(dia)
+add_subdirectory(kokoro)
+add_subdirectory(orpheus)
+add_subdirectory(parler)
