+++ src/models/dia/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(tts PRIVATE
+        model.cpp
+        model.h
+)
