+++ src/models/kokoro/CMakeLists.txt
@@ -1,0 +1,6 @@
+target_sources(tts PRIVATE
+        model.cpp
+        model.h
+        phonemizer.cpp
+        phonemizer.h
+)
