+++ src/models/parler/CMakeLists.txt
@@ -1,0 +1,5 @@
+target_sources(tts PRIVATE
+        model.cpp
+        model.h
+)
+add_subdirectory(t5)
