+++ src/models/parler/t5/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(tts PRIVATE
+        model.cpp
+        model.h
+)
